# rgw/beast: Enable SSL session-id reuse speedup mechanism